backend/requirements.txt
fastapi
uvicorn
pillow
pytesseract
python-multipart
requests
google-generativeai
spacy